+++ src/core/CanSignalTrendData.h
@@ -15,7 +15,7 @@ public:
     const QVector<TrendPoint> &points() const { return m_pts; }
     double minValue() const;
     double maxValue() const;
-    int    size()     const { return m_pts.size(); }
+    int    size()     const { return static_cast<int>(m_pts.size()); }
     bool   isEmpty()  const { return m_pts.isEmpty(); }
 
 private:
